back/python_requirements.txt
Flask~=2.0.1
Flask-Cors~=3.0.10
openpyxl~=3.0.9
pandas~=1.4.4
plotly~=5.10.0
pydantic~=1.10.2
PyYAML~=5.4.1
xlrd~=2.0.1
xlsxWriter~=3.0.1
